+++ CHANGELOG.md
@@ -7,23 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- React `key` for SocialNetworks and AcceptedPaymentMethods.
+
 ## [2.11.0] - 2019-04-30
+
 ### Added
+
 - Added social network name as a modifier to their CSS classes.
 - Added `title` prop to SocialNetworks block.
 
 ## [2.10.0] - 2019-04-24
+
 ### Changes
+
 - Scopes messages by domain
 
 ## [2.9.2] - 2019-04-22
 
 ## [2.9.1] - 2019-04-22
+
 ### Fixed
+
 - Build passes in new react3.x builder
 
 ## [2.9.0] - 2019-04-12
+
 ### Added
+
 - Added permission for `flex-layout` on the footer.
 
 ## [2.8.0] - 2019-04-11